+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 Version 7.1-SNAPSHOT
 -------------
 
+ADDED:
+
+- new interface for encrypt an generic java object to a File object
+- new interface for decrypt an encrypted {@link File} object that was previously encrypted and return the decrypted result as generic object
+
 Version 7
 -------------